+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/KotlinExtension.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 DiffPlug
+ * Copyright 2016-2020 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -33,8 +33,8 @@ import com.diffplug.spotless.kotlin.KtfmtStep;
 public class KotlinExtension extends FormatExtension implements HasBuiltinDelimiterForLicense {
 	static final String NAME = "kotlin";
 
-	public KotlinExtension(SpotlessExtension rootExtension) {
-		super(rootExtension);
+	public KotlinExtension(SpotlessExtension spotless) {
+		super(spotless);
 	}
 
 	@Override
@@ -76,7 +76,7 @@ public class KotlinExtension extends FormatExtension implements HasBuiltinDelimi
 		}
 
 		private FormatterStep createStep() {
-			return KtLintStep.create(version, GradleProvisioner.fromProject(getProject()), userData);
+			return KtLintStep.create(version, provisioner(), userData);
 		}
 	}
 
@@ -103,7 +103,7 @@ public class KotlinExtension extends FormatExtension implements HasBuiltinDelimi
 		}
 
 		private FormatterStep createStep() {
-			return KtfmtStep.create(version, GradleProvisioner.fromProject(getProject()));
+			return KtfmtStep.create(version, provisioner());
 		}
 	}