+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/SpotlessPluginPreConfigAvoidance.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 DiffPlug
+ * Copyright 2016-2020 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,13 +15,16 @@
  */
 package com.diffplug.gradle.spotless;
 
-import java.util.Locale;
+import org.gradle.api.Project;
+import org.gradle.api.Task;
+import org.gradle.api.plugins.JavaBasePlugin;
+import org.gradle.util.GradleVersion;
 
-import com.diffplug.common.base.StandardSystemProperty;
+class SpotlessPluginPreConfigAvoidance {
+	static final GradleVersion CONFIG_AVOIDANCE_INTRODUCED = GradleVersion.version("4.9");
 
-interface TestFixtures {
-	boolean IS_WIN = StandardSystemProperty.OS_NAME.value().toLowerCase(Locale.US).contains("win");
-	String EXPECTED_RUN_SPOTLESS_APPLY_SUGGESTION = IS_WIN
-			? "Run 'gradlew.bat :spotlessApply' to fix these violations."
-			: "Run './gradlew :spotlessApply' to fix these violations.";
+	static void enforceCheck(SpotlessExtension extension, Project project) {
+		Task check = project.getTasks().getByName(JavaBasePlugin.CHECK_TASK_NAME);
+		check.dependsOn(extension.rootCheckTask);
+	}
 }