+++ plugin-gradle/src/test/java/com/diffplug/gradle/spotless/EncodingTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 DiffPlug
+ * Copyright 2016-2020 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,7 +19,7 @@ import java.nio.charset.Charset;
 
 import org.junit.Test;
 
-public class EncodingTest extends GradleIntegrationTest {
+public class EncodingTest extends GradleIntegrationHarness {
 	@Test
 	public void defaultIsUtf8() throws Exception {
 		setFile("build.gradle").toLines(