+++ plugin-gradle/src/test/java/com/diffplug/gradle/spotless/ErrorShouldRethrowJre11Test.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 DiffPlug
+ * Copyright 2016-2020 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -32,7 +32,7 @@ import com.diffplug.spotless.JreVersion;
 import com.diffplug.spotless.LineEnding;
 
 /** Tests the desired behavior from https://github.com/diffplug/spotless/issues/46. */
-public class ErrorShouldRethrowJre8 extends GradleIntegrationTest {
+public class ErrorShouldRethrowJre11Test extends GradleIntegrationHarness {
 	private void writeBuild(String... toInsert) throws IOException {
 		List<String> lines = new ArrayList<>();
 		lines.add("plugins {");
@@ -58,7 +58,7 @@ public class ErrorShouldRethrowJre8 extends GradleIntegrationTest {
 				"    } // format",
 				"}     // spotless");
 		setFile("README.md").toContent("This code is fun.");
-		runWithSuccess(":spotlessMisc");
+		runWithSuccess("> Task :spotlessMisc");
 	}
 
 	@Test
@@ -68,9 +68,10 @@ public class ErrorShouldRethrowJre8 extends GradleIntegrationTest {
 				"}     // spotless");
 		setFile("README.md").toContent("This code is fubar.");
 		runWithFailure(
-				":spotlessMiscStep 'no swearing' found problem in 'README.md':",
-				"No swearing!",
-				"java.lang.RuntimeException: No swearing!");
+				"> Task :spotlessMisc FAILED\n" +
+						"Step 'no swearing' found problem in 'README.md':\n" +
+						"No swearing!\n" +
+						"java.lang.RuntimeException: No swearing!\n");
 	}
 
 	@Test
@@ -80,7 +81,7 @@ public class ErrorShouldRethrowJre8 extends GradleIntegrationTest {
 				"    enforceCheck false",
 				"}     // spotless");
 		setFile("README.md").toContent("This code is fubar.");
-		runWithSuccess(":compileJava UP-TO-DATE");
+		runWithSuccess("> Task :compileJava NO-SOURCE");
 	}
 
 	@Test
@@ -90,7 +91,7 @@ public class ErrorShouldRethrowJre8 extends GradleIntegrationTest {
 				"    } // format",
 				"}     // spotless");
 		setFile("README.md").toContent("This code is fubar.");
-		runWithSuccess(":spotlessMisc",
+		runWithSuccess("> Task :spotlessMisc\n" +
 				"Unable to apply step 'no swearing' to 'README.md'");
 	}
 
@@ -101,7 +102,7 @@ public class ErrorShouldRethrowJre8 extends GradleIntegrationTest {
 				"    } // format",
 				"}     // spotless");
 		setFile("README.md").toContent("This code is fubar.");
-		runWithSuccess(":spotlessMisc",
+		runWithSuccess("> Task :spotlessMisc",
 				"Unable to apply step 'no swearing' to 'README.md'");
 	}
 
@@ -113,14 +114,14 @@ public class ErrorShouldRethrowJre8 extends GradleIntegrationTest {
 				"    } // format",
 				"}     // spotless");
 		setFile("README.md").toContent("This code is fubar.");
-		runWithFailure(
-				":spotlessMiscStep 'no swearing' found problem in 'README.md':",
-				"No swearing!",
-				"java.lang.RuntimeException: No swearing!");
+		runWithFailure("> Task :spotlessMisc FAILED\n" +
+				"Step 'no swearing' found problem in 'README.md':\n" +
+				"No swearing!\n" +
+				"java.lang.RuntimeException: No swearing!\n");
 	}
 
 	private void runWithSuccess(String... messages) throws Exception {
-		if (JreVersion.thisVm() != JreVersion._8) {
+		if (JreVersion.thisVm() != JreVersion._11) {
 			return;
 		}
 		BuildResult result = gradleRunner().withArguments("check").build();
@@ -128,7 +129,7 @@ public class ErrorShouldRethrowJre8 extends GradleIntegrationTest {
 	}
 
 	private void runWithFailure(String... messages) throws Exception {
-		if (JreVersion.thisVm() != JreVersion._8) {
+		if (JreVersion.thisVm() != JreVersion._11) {
 			return;
 		}
 		BuildResult result = gradleRunner().withArguments("check").buildAndFail();
@@ -138,10 +139,10 @@ public class ErrorShouldRethrowJre8 extends GradleIntegrationTest {
 	private void assertResultAndMessages(BuildResult result, TaskOutcome outcome, String... messages) {
 		String expectedToStartWith = StringPrinter.buildStringFromLines(messages).trim();
 		int numNewlines = CharMatcher.is('\n').countIn(expectedToStartWith);
-		List<String> actualLines = Splitter.on('\n').splitToList(LineEnding.toUnix(result.getOutput()));
+		List<String> actualLines = Splitter.on('\n').splitToList(LineEnding.toUnix(result.getOutput().trim()));
 		String actualStart = String.join("\n", actualLines.subList(0, numNewlines + 1));
 		Assertions.assertThat(actualStart).isEqualTo(expectedToStartWith);
-		Assertions.assertThat(result.tasks(outcome).size() + result.tasks(TaskOutcome.UP_TO_DATE).size())
+		Assertions.assertThat(result.tasks(outcome).size() + result.tasks(TaskOutcome.UP_TO_DATE).size() + result.tasks(TaskOutcome.NO_SOURCE).size())
 				.isEqualTo(result.getTasks().size());
 	}
 }