+++ plugin-gradle/src/test/java/com/diffplug/gradle/spotless/GradleIncrementalResolutionTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 DiffPlug
+ * Copyright 2016-2020 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -28,7 +28,7 @@ import org.junit.Test;
 import com.diffplug.common.base.Errors;
 import com.diffplug.common.base.StringPrinter;
 
-public class GradleIncrementalResolutionTest extends GradleIntegrationTest {
+public class GradleIncrementalResolutionTest extends GradleIntegrationHarness {
 	@Test
 	public void failureDoesntTriggerAll() throws IOException {
 		setFile("build.gradle").toLines(