+++ plugin-gradle/src/test/java/com/diffplug/gradle/spotless/LicenseHeaderTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 DiffPlug
+ * Copyright 2016-2020 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,7 +21,7 @@ import java.time.YearMonth;
 import org.eclipse.jgit.api.Git;
 import org.junit.Test;
 
-public class LicenseHeaderTest extends GradleIntegrationTest {
+public class LicenseHeaderTest extends GradleIntegrationHarness {
 	private static final String NOW = String.valueOf(YearMonth.now().getYear());
 
 	private static final String TEST_JAVA = "src/main/java/pkg/Test.java";