+++ plugin-maven/src/test/java/com/diffplug/spotless/maven/MultiModuleProjectTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 DiffPlug
+ * Copyright 2016-2020 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,11 +15,11 @@
  */
 package com.diffplug.spotless.maven;
 
-import static com.diffplug.spotless.maven.MavenIntegrationTest.SubProjectFile.file;
+import static com.diffplug.spotless.maven.MavenIntegrationHarness.SubProjectFile.file;
 
 import org.junit.Test;
 
-public class MultiModuleProjectTest extends MavenIntegrationTest {
+public class MultiModuleProjectTest extends MavenIntegrationHarness {
 
 	@Test
 	public void testConfigurationDependency() throws Exception {